requirements.txt
appdirs==1.4.3
CacheControl==0.12.6
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
html5lib==1.0.1
idna==3.2
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
py4j==0.10.7
pyparsing==2.4.6
pyspark==2.4.5
python-dateutil==2.8.2
pytoml==0.1.21
requests==2.26.0
retrying==1.3.3
six==1.14.0
urllib3==1.26.6
webencodings==0.5.1
